requirements.txt
# OfficeMind — Python Dependencies
# Install: pip3 install -r requirements.txt

# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ── LLM & Agent ──────────────────────────────────────────────────────────────
openai>=1.40.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0

# ── RAG Pipeline ─────────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
tiktoken>=0.7.0

# ── Vision & OCR ─────────────────────────────────────────────────────────────
pillow>=10.4.0
imagehash>=4.3.1
paddlepaddle>=2.6.0; platform_machine != "aarch64"
# For aarch64 (DGX Spark): use paddlepaddle-gpu from NVIDIA wheel
# pip install paddlepaddle-gpu --extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu123/

# ── Browser Automation ───────────────────────────────────────────────────────
playwright>=1.45.0
aiohttp>=3.10.0

# ── Document Processing ──────────────────────────────────────────────────────
pypdf>=4.3.0
python-docx>=1.1.2
openpyxl>=3.1.5
pandas>=2.2.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.5.0
rich>=13.7.0
pytest>=8.3.0
pytest-asyncio>=0.23.0
